Add an interface for other MIDI controllers
