+++ go.mod
@@ -1,5 +1,5 @@
 module github.com/jmacd/launchmidi
 
-go 1.18
+go 1.21
 
 require gitlab.com/gomidi/midi/v2 v2.0.25
